# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hyperhire Tinder API",
        "description": "API Documentation for Hyperhire Tinder Application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server V1"
        }
    ],
    "paths": {
        "/people/{user_id}/like": {
            "post": {
                "tags": [
                    "People"
                ],
                "summary": "Like a person",
                "description": "Like a person by their user ID",
                "operationId": "5080c7339e18376e4bf0a25efd5e724a",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID of the user to like",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "X-Device-ID",
                        "in": "header",
                        "description": "Device ID for authentication",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "device-12345"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Person liked successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Person liked."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "X-Device-ID header is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/people/{user_id}/dislike": {
            "post": {
                "tags": [
                    "People"
                ],
                "summary": "Dislike a person",
                "description": "Dislike a person by their user ID",
                "operationId": "e0d47b8795e2b86880e4ed54b466d8d1",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID of the user to dislike",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "X-Device-ID",
                        "in": "header",
                        "description": "Device ID for authentication",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "device-12345"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Person disliked successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Person disliked."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "X-Device-ID header is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/people/recommended": {
            "get": {
                "tags": [
                    "People"
                ],
                "summary": "Get recommended people based on location",
                "description": "Returns a paginated list of recommended people sorted by distance from the provided coordinates",
                "operationId": "648fbff13f772eeb614327986a34f2f0",
                "parameters": [
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Latitude for recommendation",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": -6.2088
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "description": "Longitude for recommendation",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 106.8456
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "example": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Alicia"
                                                    },
                                                    "age": {
                                                        "type": "integer",
                                                        "example": 23
                                                    },
                                                    "pictures": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        },
                                                        "example": [
                                                            "https://cdn.app.com/1.jpg",
                                                            "https://cdn.app.com/2.jpg"
                                                        ]
                                                    },
                                                    "distance_km": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 1.2
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The lat field is required."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "People",
            "description": "People"
        }
    ]
}